requirements.txt
gitdb2==2.0.5
GitPython==3.0.2
smmap2==2.0.5
